backend/requirements.txt
Django==5.2.7
psycopg2-binary==2.9.10
gunicorn>=20.1
django-environ>=0.9
django-cors-headers==4.6.0
djangorestframework==3.15.2
requests==2.32.3
reportlab==4.0.7
openpyxl==3.1.2
